+++ web/modules/custom/qs_themes/src/Form/FilterForm.php
@@ -5,10 +5,10 @@ namespace Drupal\qs_themes\Form;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\Http\RequestStack;
 use Drupal\Core\Language\LanguageManagerInterface;
 use Drupal\Core\Session\AccountProxyInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\HttpFoundation\RequestStack;
 
 /**
  * A theme filter form that add the chosen theme as GET param to the route.
@@ -32,7 +32,7 @@ class FilterForm extends FormBase {
   /**
    * Request stack that controls the lifecycle of requests.
    *
-   * @var \Symfony\Component\HttpFoundation\RequestStack
+   * @var \Drupal\Core\Http\RequestStack
    */
   protected $requestStack;
 
@@ -63,7 +63,7 @@ class FilterForm extends FormBase {
     $form['#method'] = 'GET';
 
     // The request should be took at the last moment, avoid it on constructor.
-    $master_request = $this->requestStack->getMasterRequest();
+    $master_request = $this->requestStack->getMainRequest();
 
     // Disable caching & HTML5 validation.
     $form['#cache']['max-age'] = 0;